requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<3.0.0
Pillow>=9.0.0
scipy>=1.9.0
# Advanced features - now fully supported
tensorflow>=2.13.0
face-recognition>=1.3.0
dlib>=19.24.0
pygame>=2.5.0
sounddevice>=0.4.0
# Performance monitoring and enhancement
psutil>=5.9.0
# Data visualization and dashboard
matplotlib>=3.7.0
